requirements.txt
appdirs==1.4.3
bottle==0.12.13
cycler==0.10.0
matplotlib==2.0.0
neat-python==0.91
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.7
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
